# swagger-utf16.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "E.L.F Brewery API",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/v1/breweries/autocomplete":  {
                                                         "get":  {
                                                                     "tags":  [
                                                                                  "Breweries"
                                                                              ],
                                                                     "parameters":  [
                                                                                        {
                                                                                            "name":  "query",
                                                                                            "in":  "query",
                                                                                            "schema":  {
                                                                                                           "type":  "string"
                                                                                                       }
                                                                                        }
                                                                                    ],
                                                                     "responses":  {
                                                                                       "200":  {
                                                                                                   "description":  "Success"
                                                                                               }
                                                                                   }
                                                                 }
                                                     },
                  "/api/v1/breweries/{Name}":  {
                                                   "get":  {
                                                               "tags":  [
                                                                            "Breweries"
                                                                        ],
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "name",
                                                                                      "in":  "path",
                                                                                      "required":  true,
                                                                                      "schema":  {
                                                                                                     "type":  "string"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "Success"
                                                                                         }
                                                                             }
                                                           }
                                               },
...
              },
    "components":  {
...
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "apiKey",
                                                              "description":  "Enter \u0027Bearer\u0027 followed by your JWT token",
                                                              "name":  "Authorization",
                                                              "in":  "header"
                                                          }
                                           }
                   },
    "security":  [
                     {
                         "Bearer":  [

                                    ]
                     }
                 ]
}